# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hill Wood",
        "description": "Hill Wood App API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "Hill Wood API Server"
        }
    ],
    "paths": {
...
        "/api/Subscribe/Channel/{id}": {
            "post": {
                "tags": [
                    "subscriber"
                ],
                "summary": "subscribe new subscriber",
                "description": "Logged in user can subscribe new channel",
                "operationId": "subscribeChannel",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Channel id you need to subscribe",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad user Input"
                    },
                    "200": {
                        "description": "Operation ok"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/Unsubscribe/Channel/{id}": {
            "post": {
                "tags": [
                    "subscriber"
                ],
                "summary": "Unsubscriber channel",
                "description": "Logged in user can unsubscribe any chaannel he/she subscribed",
                "operationId": "unsubscribeChannel",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Channel id you need to unsubscribe",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad user Input"
                    },
                    "200": {
                        "description": "Operation ok"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/Channel/Subscribers/{id}": {
            "get": {
                "tags": [
                    "subscriber"
                ],
                "summary": "get channel subscribers",
                "description": "Get any Channel subscribers",
                "operationId": "channelSubscribers",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Channel id you need to show subscribers.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "204": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad user Input"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}